backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
nltk
pandas
numpy
textblob
beautifulsoup4
requests
motor
passlib[bcrypt]
passlib[argon2]
PyJWT
python-dotenv
pydantic-settings
python-multipart
langdetect
deep-translator
googlesearch-python
bleach
email-validator
dotenv
google-generativeai
cryptography
redis[hiredis]
celery[redis]
prometheus-client
sentry-sdk[fastapi]